+++ srcs/ClassSchedule.Infrastructure/IServiceCollectionExtensions.cs
@@ -1,4 +1,5 @@
 using Microsoft.Extensions.DependencyInjection;
+using System;
 
 namespace ClassSchedule.Infrastructure;
 
@@ -19,7 +20,9 @@ public static class IServiceCollectionExtensions
         /// <returns>服务集合</returns>
         public IServiceCollection AddInfrastructure()
         {
-            return services.AddSingleton<FileLoggerOptions>();
+            return services
+                .AddSingleton(TimeProvider.System)
+                .AddSingleton<FileLoggerOptions>();
         }
     }
 }